+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.0.12] - 2026-06-14
+
 ### Changed
 
 - `Context.derive()` no longer eagerly copies all parent variables into every child